FastAPI/requirements.txt
fastapi[standard]
sqlalchemy
python-dotenv
psycopg2